Moved functionality to pxTableData
